voting and ranking adaptation to 10 second tmer now running in frontend
event names according to new specification (picture on whatsapp)
Game now has to be stored in gamecontroller since the sequence of events is no longer automatic
Simplified timer to one instance that can be used serveral times. Duration is set as parameter, this makes it easier to read and use. Esp since there is one timer that had not yet been defined.
send ranking as hashmap